Pushing my changes from james-wsgi branch onto the trunk.
Biomart doesn't work on the trunk yet (Before or after this commit).

axt to lav tool won't be functional until multiple datasets per history item is available on trunk - although I could
make it so that only the lav file is generated and no fasta files, in which case it wouldn't be an issue.