Use absolute paths for references to datasets and associated files.
NOTE:

Our cluster shouldn't have a problem with this since datasets live in the same
place on the web server and computer nodes. However for the long term we need
a better way to deal with this. Different compute nodes may have files in
different locations, a more complicated resolution strategy needs to be
developed.