Streamlined interface for supporting new data types - requires db change and note config change. db change: update dataset set extension = 'txt' where extension = 'text';
Adding support for a new data type is now as simple as adding the new type to the config file and optionally adding the
type into the new sniff_order section in the config file, and adding the new class for the data type to the appropriate
datatypes code.

The config file now has a new section [galaxy:sniff_order] which lists the order in which datatype class sniff functions will
be called (each datatype class can now optionally include its own sniff function, although not a requirement).

Support for the 'text' file extension has been eliminated as we will standardize on the 'txt' file extension for the Text data type.