Initial check in of library functionality.
The ui is Very rough, at best.

Libraries can be accessed by users through a new tool under datasources.

Libraries can be created and accessed for modification via a link on the admin page.

Currently, all library datasets are currently associated with the public group.

Tags have been defined (tables and mappings) and can be associated with a dataset or a folder, but there is currenty no interface for displaying or creating these.

Creation and editing of a library should be an admin only function, however this check is not currently present.

The creation/editing page for a dataset is currently hard coded, it should be dynamic (i.e. datatypes and dbkeys)

Brief description of a library:
A library is composed of a name, description and a root folder.
A folder is composed of a name, description, and a parent folder (none when is the root), where a folder can have many or none child folders and datasets.
A library dataset is composed of most of the same attributes as a history dataset, but it is linked via a folder_id to a folder (instead of a history_id to a history).

Ordering of the library is maintained by the 'order_id' attribute of each folder and dataset object (which is populated by the item_count attribute of a folder)