Add the ability for metadata to be set in an external process, this requires set_metadata_externally = True in universe_wsgi.ini.
Currently this is supported in the local runner (long running set_meta() can now be canceled the same as with jobs) and in the non-staging PBS runner.

There is a slight change to how new Metadata FileParameters are to be created in datatype.set_meta(), see the MAF datatype.

TODO:
Prevent editing of metadata on a running job when metadata is to be set externally.
Have datasets which were copied before job completion use the metadata from the original output dataset; currently .set_meta() is called locally on these.