Fixes and functional tests for uploading directories of files from both the Admin view and the Libraries view ( requires new config setting ).
In addition to the config setting "library_import_dir" for the Admin view, the new config setting "user_library_import_dir" for the Libraries view will allow non-amin users to upload a directory of files.

The configured directory must contain sub-directories named the same as the non-admin user's Galaxy login ( email ).  The non-admin user is restricted to uploading files or sub-directories of files contained in their directoy.