Add a checkbox to the Create Group page that if checked will create a new Role with the same name.  This provides a similar feature the the existing checkbox on the Create Role page.
However, the behavior is now changed such that new associations are created when the checkbox is checked whereas before, only the Group or Role objects with the same name were created, but not associated with anything.

This code was very old, so I did some cleanup / improvements as well.  Functional tests enhanced to reflect changes.