Add more featureful job limiting and optimize the query for checking whether jobs are ready to run.  Input dependency checks are now performed via SQL rather than walking up the object chain.  Limits on the number of jobs a user can run can now be set across the entire instance and a job runner URL.
Quota checks at job runtime are only performed once, after limit checks.  If a
user is over quota, jobs are moved to a "paused" state.  Once the user is under
quota, jobs can be unpaused and continue to run (once this UI is added in
another commit, shortly).  This obviates the need for quota checks on every
job, on every queue cycle.

When a job's input dataset errors, the job (and all jobs dependent upon that
job) are no longer errored.  It will then be possible to remap a job to new
inputs to allow execution to continue from the point of failure.  Commit for
that is also coming shortly.