FASTQ Opt: In convert_read_to_format, adjust when new_encoding logic is calculated. Doesn't change the behavior or optimize anything, this is just done to simplify subsequent commits.
Baseline:

I took the first 100 megabytes of a 26 gigabyte example of a FASTQ file filtered with the FASTQ filter tool that I found in an "Important Galaxy User"'s history on main. I ran with the same command-line on my dev box with the start of this file and using Python's -m to profile function times and total amount of time to serve as a baseline as I optimized the FASTQ filter code. Here is the start of the output:

Kept 135848 of 250000 reads (54.34%).
         200015991 function calls (199765991 primitive calls) in 136.934 seconds

Extrapolating this out, that 26 gigabyte file would take roughly 10 hours to process on my laptop - this is slightly longer than what it took on main - indicating to me this is likely not disk bound since my SSD would probably outperform main?