Fix small bug in handling of 'brand' text.
This was changed in 5828950834e3c888683dbeb567e10bcaf335c94f, after that commit the brand was no longer preceeded by a / in the masthead the way it is in the title. I assume this was an oversite and not intentional, if intentional feel free to backout of this commit.