Update LWR client through LWR changeset 5c609be. The LWR server and client can be driven via message queues, though this changeset doesn't update the LWR runner itself to allow this - only the client itself. The runner however has been updated to support some more minor changes that enabled the LWR client to be message queue driven.
In particular, a new action type was added "remote_copy" that does a file-system copy of files being staged in or out like "copy" - but it does so on the LWR side instead of the Galaxy client. In this scenario these actions are serialized and transferred with the "launch" command to the LWR - potentially eliminating many HTTP calls.

Paired with this, more processing can happen on the client side - namely calculating remote paths and properties - and if configured in this fashion the LWR "setup" step can be eliminated - allowing the LWR client to completely specify a job with a single one-directional message (very useful for message queues). While modifying the LWR to be able to behave in this fashion was done to enable message queues - a traditional HTTP driven LWR can use these features (remote copy, client-side path calculation, etc...) and job_conf.xml.sample_advanced has been updated to reflect this.

For more information on changes to the LWR client this encompasses see the following LWR changesets:

https://bitbucket.org/jmchilton/lwr/commits/bacaa5c840bbf41952fcafe6e0deadd306c868fa
https://bitbucket.org/jmchilton/lwr/commits/c8254d4525fa19f6761c6a75506d7b2bbdbced9e
https://bitbucket.org/jmchilton/lwr/commits/1f068927a677bbfb932c475e5a4dad9885c61071
https://bitbucket.org/jmchilton/lwr/commits/bde8cbe6f5736a82d5fa20b322cb34d9d4e8fbee
https://bitbucket.org/jmchilton/lwr/commits/993910fa0c5a0ac2be72b2d70eeaaa0767e18032