Rearrange auth to use a Galaxy manager object pattern.
On the positive-side this feels much more like a traditional Galaxy style access pattern and each operation is likely much more efficient since the plugins aren't be reloaded and XML re-parsed for each action. The downside is a Galaxy restart is required to update the authentication providers for Galaxy - while this is less than ideal in some ways - it is what Galaxy admins would expect for the other configuration files anyway so it is arguably better to not be more clever here.