Simply external set_metadata's requirements.
Do not pass the config, files/temp, or galaxy root paths to the script.
Instead, directly pass the path(s) to the files on which we are setting
metadata, and determine the galaxy_root directory automatically based on
the location of the set_metadata script.

This is a fix to the previous commit for allowing the web and handler
processes to run from different directories.