Allow updating TS repo via API
In support of https://github.com/galaxyproject/tools-iuc/issues/73 we
would like to be able to push changes made to .shed.yml with planemo to
the toolshed.

This extracts the logic (mostly copy+paste) from the original repository
controller into the repository util, then wraps that in an API PUT
method.