Allow testing tool outputs based on MD5 hashes.
This should not be considered the best practice for most text based files, but for binary files we cannot produce diffs anyway so this is probably better than storing the pointless binary blobs in repositories for instance. In particular Dan and I thought this would be useful for testing index files for data managers.

Includes test tool demonstrating this functionality, run test with

    ./run_tests.sh -framework -id md5sum