ScatterPlot No API connection for nested www_root
ScatterPlot makes use of a hardcoded API url which is incorrect for certain server configuratoins. Currently these servers need to be manually configured to link to the correct API URL. This patch creates the correct URL from the parental variable parent.galaxy_config.root and returns "/" if this variable can not be found. More visualisations suffer from this bug and have to be patched as well.