Remove broken method from library_common.
That method could never have done anything but throw an exception or return [], this commit just assumes that it was always returning [] and updates the one place it was called accordingly. Not sure what widgets are - but I don't think @martenson is implementing them in the data library rewrite.

This method was originally added in f7334104a000f36fd1c5f8f18cf617b399b7ff0f