Implement policies for commmitters and pull requests.
- State and clarfiy my (@jmchilton's) understanding of existing informal pull request policies.
- Add a list of committers who should have "commit" access to the repository according to my understanding of the team meeting held on 5/28/2015.
- Add language describing how the Galaxy project might manage commit access and project governance going forward in accordance with my understanding of the team meeting discussion.

This fills in a lot of gaps left by informal policies and informal discussions and I am eager to see much discussion on each of these topics. Much of the language and structure is borrowed with heavy modification from the Django and Apache variants of these documents.

Also - congrats to @nsoranzo, @erasche, and @bgruening who the team unanimously voted to grant commit access to.