Fix sample datatype definition for CSV.
For me it seems like the viz plugins only detect the datatype if the plugin datatype (e.g. ``tabular.CSV``) is explicitly instantiated by some extension directly. In your sample config you have ``subclass="True"`` so ``tabular.CSV`` wasn't working. This is probably a bug with the datatype code - but nonetheless I think dropping subclass from the sample is an okay workaround - because it does seem to be a direct mapping.