Fix wrong call to LDAPObject.result() by using search_ext_s().
result() second parameter is 'all', which should be 0 or 1.
result() call was passing 60 as second parameter, which was probably
intended to be a timeout (third parameter).
By using search_ext_s(), the search() and result() calls are combined in one
and the number of search results retrieved is limited to 1.

Also, if the search return no results, this is now logged separately at
warning level.