Fix workflow run for certain kinds of explicit output collections.
Workflows with steps that produce an unknown number of datasets must explicitly be sent to the new workflow engine - since the workflow needs be re-evaluated after such steps have completed. This change updates the force_queue to check workflow steps for such collections and return True if any are found - this is what causes certain GUI driven workflows to use the new workflow engine.