Fix for missing included files on history export and import.
Notes:

If a history was exported previously, galaxy would just serve the cached version.
In that case, it is better to make some changes to the history for example, by re-running
the last job and then exporting it.

Export of histories that haven't been exported previously should work without problems.