Fix for server error when installing some repos.
Examples: IUC's gatk2 repository, devteam's samtools_rmdup.

When attempting to install either of those, or another repository
that has tool tests with an input parameter that does not have a
value attribute, the repository would be stuck in state 'Cloning'
and show a debug link in the log. This commit resolves that error.