Add dynamic_proxy_prefix configuration
Normally the proxy binds to :8800 and runs containers under
:8800/ipython/... which is great... until you have an upstream proxy
like apache wrapping that at the url FQDN/ipython and suddenly your
cookies aren't available because they're specific to /galaxy.

Thus, when an upstream proxy is in use, we correct the proxy's path to:
with :8800/{cookie_path}/gie_proxy/ipython/... which behaves much more
nicely with upstraem proxies.