Try harder to verify script integrity before running job scripts.
These "Text file busy" errors are occasionally encountered in the wild in production settings and more commonly when running Galaxy in Docker containers.

Simply sycning the file system has proven insufficient to prevent the problem and I do not want to add an arbitrary sleep. So this new strategy is used.

 - Inject special functionality into each job script that when executed with a certain environment variable set causes the script to exit with a specific return code.
 - Have Galaxy rerun this script in that configuration until it can succeed in executing the script.
 - In between each iteration sleep for a small amount of time and try to cause a file system sync.

This entire is ... insane ... so I have added a configuration option to disable it entirely check_job_script_integrity=False.