Fix API installation of tool_dependencies
This should fix an installation issue with tool_dependencies via the API.
Tool dependency repositories to not have any tools specified and with this no metadata associated.

It was broken in https://github.com/galaxyproject/galaxy/commit/bb6a784f13ff3f91762244d4488437ffef9e5acf
So I think this should be part of 15.10.

Thanks to Greg for fixing this one.