Modify function "pretty_print_time_interval"
I've been having an issues with the API when creating data libraries through the new interface. I have a Galaxy server running on Ubuntu 14.04.3 with a MariaDB server in the back. Apparently, the time returned from the database doesn't include the milliseconds, hence dropping the requirement for milliseconds fixes my issues.
I've been in talks with @martenson and @natefoo about this over at Biostar
https://biostar.usegalaxy.org/p/15330/#15373