Move Dictifiable out of model package into util.
Many objects that aren't model objects are Dictifiable at this point so there is no reason to have this in model. This changes eliminates Dictifiable's dependency on sqlalchemy (by using a more pythonic type check). Now that Dictifiable has no external dependencies, it is much easier to reuse it (and more important objects that extend it) in other projects such as planemo.