[WIP] Make more job options configurable at the destination level.
Makes a host of job-related parameters configurable at the destination level instead of the of only at the app level. This solves major problems like allowing the local job runner and the drmma run as real-user runner to operate within in the same environment and provides little conveniences like allowing configuration of cleanup job at the destination level.