Give tool executions a clean working directory.
Rather then massively renaming job_directory and job_working_directory, I introduce the concept of the tool execution working directory (tool_working_directory) that is the job working directory + "/working" and that will not be populated with job files.

Some legacy cruft in there for the upgrade, things like galaxy.json and dynamic outputs may be in the job working directory or the tool execution working directory.