Implement new GIE proxy
This provides an alternative proxy to the original NodeJS proxy.

The implementation is done in golang, which has the nice property of compiling
to a single binary. While it's a negative in terms of number of devteam who
write that language, I'd argue it's not that much worse than nodejs in terms of
accessibility.

Features include:

- Simple rest-ish API instead of file-based IPC
- Automatic culling of containers, no more monitor_traffic.sh hacks
- Admin configurable, cross GIE timeout setting