Work toward automating deployment.
Create new makefile targets for release management.

 - ``release-create-issue``: Automates creating a release checklist - https://github.com/jmchilton/galaxy/issues/41)
   Check list contains actual tailored ``make`` commands and links (for PRs and issues).
 - ``release-check-metadata``: Checks milestone PRs for required metadata needed for generating bootstrapped release notes.
 ` ``release-bootstrap-history``: Bootstrap release notes from PRs and metadata.
 - ``release-check-blocking-prs``: Check for PRs blocking release, should happen before branching release and before tagging release.
 - ``release-check-blocking-issues``: Check for Issues blocking release, should happen before tagging release. (Only allowed issue is the release creation issue.)

The issue created with ``release-create-issue`` documents how and when to execute subsequent ``make`` targets (including the target to create the issue for the next release).

This PR contains a bunch of other enhancements to the bootstrap_history.py script to enable these targets and renames ``create_release_rc`` Makefile target so all release related targets start with the prefix ``release-``.

Conflicts:
	Makefile