Adjust testing directories so no python root is ever "tool_shed".
We have a tool_shed directory in lib, so if we enter a test case from some path and test/tool_shed happens to be a valid python top-level module - it blocks out lib/tool_shed and Galaxy doesn't work properly in the test.

A similar problem was encountered with unit tests and fixed with https://github.com/galaxyproject/galaxy/commit/df442a06382f23bdf0401bdd632b9584d3431281 - but I have decided that directory name was a bit log and so this commit unifies the approach to unit and functional test names for tool shed tests.