UI, history: correct current dataset, tags, annos
Correct the 'currently used dataset' indicator (the left, blue border)
making sure it clears when non-dataset related content loads in the
center panel.

Correct the toggling of tags and annotations in the current history
view forcing the subviews to show/hide based on the history view
and not by toggling them individually.