Build tool lineage for versionless toolshed tool_ids
The tool_id for a toolshed tool contains the version as last element
(<toolshed>/<name>/<owner>/<id>/<version>). This means that looking up a tool
lineage using a tools' tool_id does not work if the tool is installed in a different
version than referenced by the tool_id. To allow this, we register the
tool_lineage also under a tool_id that is stripped of the version.