Automate and clarify more of the release process.
upstream is used repeatedly and just assumed to be a remote - this tests for it and adds it if it doesn't exist before using in make commands now. This also automates the previous release to dev merge in the instructions with a new Make command.

Additionally, two PRs need to be created to update versions. This prints a reminder to do this and adds it to the release checklist. Ideally, these PRs could be avoided with a direct push or automatically created with hub (https://github.com/github/hub).