API: add server_starttime to api/configuration
Adds key server_starttime to the JSON returned from api/configuration.
This is the epoch time since the last server restart and may be
convenient for use in client-side caching.

This moves server_starttime from web/framework/helpers/__init__.py
(for use in script cache-busting) to web/framework/base.py.

On the client, this should be available to all pages as:
Galaxy.config.server_starttime.