UI, display_base: render links in header first
This allows text to flow around them and for them to be positioned at/
near the top (and remain clickable).

This is a stop-gap/kludge and what we'd really want to do is move away
from absolutely positoned panels/layouts so elements like this can
stretch more naturally around their contents.

Fixes #2172