Generate metadata in separate conda environment
Galaxy supports generating metadata in a separate command (enable
beta_tool_command_isolation). This command has its own dependency resolution,
that may interfere with thedependency resolution of the actual tool, if the
same environment is being used. With this PR we create a new environment when
resolving metadata dependencies with conda, that will be active only for
metadata generation.