Switch Dockerized commands to use sh instead of bash.
We switched to bash for the conda and that doesn't get sourced in containers - so we should be a bit pragmatic about this and switch back to /bin/sh for containers which is more common than /bin/bash.

xref https://github.com/galaxyproject/galaxy/pull/2278

Update galaxy.ini.sample to mention this as suggested by @bgruening in https://github.com/galaxyproject/galaxy/pull/2282#issuecomment-216232788.