UI, history: update using xhr date header
Makes the 'updated since' contents query based on any previous update's
ajax Date response header instead of the client time. This may offer
more protection against incorrect client time.

This works (and may work better than the previous method) because the
Date header is floored and not rounded and seems to mark a time before
the contents query is made.

Resolves #2242 (hopefully)