Revise HTCondor Docker image resolution.
Bring inline with the rest of the job runner's Docker image resolution stuff. Order of reosolution is:

 - Use destination specified override if found: docker_container_id_override
 - Use the tool specified container if not found and no override: see tool
 - Else - use docker_default_container_id.

This adjusts the documentation to reflect this and make the the parameters more consistent throughout. I've modified this so docker_image (the old name introduced by the Docker Condor PR) still works - but I have left it undocumented intentionally for consistency.