Update Galaxy testing to target latest galaxy-lib.
Incorporates https://github.com/galaxyproject/galaxy-lib/pull/10 which modifies the interface to the test output verification code to allow reuse from within Planemo in both non-tool and non-Galaxy contexts. I also think this represents a better separation of concerns.

One upshot of this is that CWL-style checksum verification is now possible from within Galaxy tools - a test case of this is included.