Add a password expiration config option
Add a "last_password_change" field to user table and user object.
Add a config option "password expiration_period" to allow admins
to require users to change passwords every so many days. On login,
check for the presence of the config option and if the user pw
has expired, redirect to the "change password" page.