History, UI: use id of displayed history in delete
- Refactor `delete current` to `delete` and require an id arg.
- Update history options menu to pass the id of the displayed/
current history to the above delete controller endpoint.

This means that the current history panel will delete the history
that's displayed (instead of potentially a different history that's
been set to the current history on the server side).

Note: this isn't an ideal refactoring but should solve the below
issue temporarily until we move more of the session state out of
API.

Closes #894