Refactor recent toolbox changes for galaxy-lib.
The abstracttoolbox class is supposed to shield Galaxy from the internals of the toolbox organization and vice versa, so having a reload method that uses Galaxy-specific queueing stuff is less than ideal. The concrete toolbox class is where I have been putting such functionality.

In addition to this being the abstract design I prefer, it is definitely needed for galaxy-lib which contains the abstracttoolbox but not the concrete toolbox or the queue worker functionality.

xref #2840