Datatypes: include base classes in search by name
Fixes #2982

Adds the base class 'images.Image' to the list of classes
returnable by 'datatypes_registry.get_datatype_class_by_name'.

Normally that method searches the registry's list of datatypes and
those are instantiable classes only. The images.Image can be
considered an abstract base class and you'd never create a dataset
of that datatype.

The method is used for getting classes for 'isinstance' tests, however,
so including base classes like 'images.Image' allow testing for all
its image subclasses. This change makes that possible and allows the
method (and its consumers) to behave in a much more predictable and
useful manner.