Fix BLAST database *.loc files inconsistency
We should have all of the ``blast*.loc`` files in the sample configuration, or none of them.

Historically Galaxy included only the nucleotide BLAST database ``blastdb.loc`` file for use with mega-blast, with ``blastdb_p.loc`` later for protein databases added as part of the BLAST+ suite (which initially was in the main Galaxy repository).

In a3018feb028658fc3a7071803b7b49cb575e7358 the mega-blast tool was moved to the Tool Shed, and the ``blastdb.loc`` entry removed. Likewise the BLAST+ tools were moved to the Tool Shed, but the ``blastdb_p.loc`` entry remains.

Later (on the Tool Shed only?) ``blastdb_d.loc``was added for protein domain databases.