Remove verbose_install_check option
This option was introduced with #2554 and #2538, in reponse to a false
positive conda install.  This would have been caught by simply checking
the exit code, which we are doing by default now. The additional
verification is too strigent and not in line with the conda resolver.
Activating this option will report perfectly well installed conda
environments as having failed and will cause them to be uninstalled.